refactor(engine): extract node failure-outcome decisions to engine/outcome.rs (audit 🔴-1, PR5)
Fifth decomposition slice of the `engine.rs` god-module. The stateless layer
that decides what happens to a failed/retryable node moves into a sibling
`engine/outcome.rs`:
- `effective_retry_policy` / `RetryDecision` / `error_is_terminal` /
  `compute_retry_decision` / `next_retry_at` — retry resolution;
- `FailureOutcome` / `classify_failure` / `apply_failure_recovery` /
  `route_failure_edges` — failure classification + recovery routing.

Unlike the earlier slices these are FREE functions, not `impl` methods, so the
move needs a re-export instead of method resolution: the moved items are bumped
to `pub(super)`, and `engine.rs` gains `use outcome::*;`. That private re-export
into the `engine` module is transitively visible to every child module's
existing `use super::*`, so the bare-name call sites in `frontier` / `resume` /
`engine.rs` / `tests` resolve with zero edits to those files.

`engine.rs`: 3,806 → 3,462 lines (13,298 → 3,462 cumulative across PR1–PR5,
-74%). Pure cut + the `mod outcome;` / `use outcome::*;` declarations + the
`pub(super)` bumps.

This is the last planned decomposition slice: the five largest subsystems
(tests, frontier executor, resume/signal, persistence, failure-outcome) are
now their own modules, leaving engine.rs as the cohesive core — the
`WorkflowEngine` struct, its builder/`with_*`/`execute_workflow` surface, the
`NodeTask`, and the smaller inline executor helpers.

Verified: `cargo fmt`/clippy `-p nebula-engine --all-features --all-targets`
clean; `cargo nextest run -p nebula-engine --all-features` → 442 passed,
0 skipped (identical set to pre-move).
